+++ .env.local.example
@@ -5,3 +5,5 @@ SG_API=
 SG_FORM=mail@domain.com
 
 GOOGLE_FORM_ID=
+
+EMAIL_SECRET=
